requirements.txt
python-dotenv
flask
flask-jwt-extended
flask-cors
Flask-SocketIO
Flask-Session
werkzeug
spotipy
gevent-websocket
cachetools